Refactor the admin for activity types in to admin namespace and scope
