+++ app/views/activity_categories/_filter.html.erb
