+++ app/views/activity_categories/_filters.html.erb
@@ -1,0 +1,5 @@
+<%= render 'filter', name: :key_stage, filters: filter.all_key_stages, selected_filters: filter.selected_key_stages, title: 'Key Stages'%>
+<%= render 'filter', name: :subject, filters: filter.all_subjects, selected_filters: filter.selected_subjects, title: 'Subjects'%>
+<%= render 'filter', name: :topic, filters: filter.all_topics, selected_filters: filter.selected_topics, title: 'Topics'%>
+<%= render 'filter', name: :impact, filters: filter.all_impacts, selected_filters: filter.selected_impacts, title: 'Impacts'%>
+<%= render 'filter', name: :activity_timing, filters: filter.all_activity_timings, selected_filters: filter.selected_activity_timings, title: 'Timings'%>
