--- app/views/activity_types/_activity_type.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! activity_type, :id, :name, :description, :created_at, :updated_at
-json.url activity_type_url(activity_type, format: :json)
