--- app/views/activity_types/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.array! @activity_types, partial: 'activity_types/activity_type', as: :activity_type
