--- app/views/activity_types/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! "activity_types/activity_type", activity_type: @activity_type
