+++ app/views/admin/activity_types/_card.html.erb
