+++ app/views/admin/activity_types/_crud_filter_checkboxes.html.erb
