+++ app/views/admin/activity_types/_filter.html.erb
@@ -1,0 +1,11 @@
+<fieldset class="form-group">
+  <legend><%= title %></legend>
+  <%= collection_check_boxes name, :"#{name}_ids", filters, :id, :name, { multiple: true } do |b| %>
+    <div class="custom-control custom-checkbox">
+      <%= b.check_box( checked: (selected_filters.include?(b.object)), class: "custom-control-input") %>
+      <%= b.label(class: "custom-control-label") do %>
+        <%= b.object.name %>
+      <% end %>
+    </div>
+  <% end %>
+</fieldset>
