+++ app/views/admin/activity_types/_filters.html.erb
