+++ app/views/admin/activity_types/index.html.erb
@@ -1,4 +1,9 @@
 <h1>Activity Types</h1>
+<div class="container">
+  <div class="row">
+    <p><%= link_to 'New Activity Type', new_admin_activity_type_path, class: 'btn' %></p>
+  </div>
+</div>
 
 <% ActivityCategory.all.each do |activity_category| %>
   <div class="container">
@@ -10,15 +15,17 @@
           <th scope="col">Name</th>
           <th scope="col">Score</th>
           <th scope="col">Key Stages</th>
+          <th scope="col"></th>
         </tr>
       </thead>
 
       <tbody>
         <% @activity_types.where(activity_category: activity_category).each do |activity_type| %>
           <tr class="<%= active(activity_type.active) %>" scope="row" >
-            <td><%= link_to activity_type.name, activity_type_path(activity_type) %></td>
+            <td><%= link_to activity_type.name, admin_activity_type_path(activity_type) %></td>
             <td><%= activity_type.score %></td>
             <td><%= activity_type.key_stage_list %></td>
+            <td><%= link_to 'Edit', edit_admin_activity_type_path(activity_type), class: 'btn btn-sm' %></td>
           </tr>
         <% end %>
       </tbody>
@@ -27,3 +34,9 @@
   </div>
 
 <% end %>
+
+<div class="container">
+  <div class="row">
+    <%= link_to 'New Activity', new_admin_activity_type_path, class: 'btn' %>
+  </div>
+</div>
