+++ app/views/admin/activity_types/new.html.erb
@@ -3,5 +3,5 @@
 <%= render 'form', activity_type: @activity_type %>
 
 <div class="other-actions">
-  <%= link_to 'Back', activity_types_path, class: 'btn btn-secondary' %>
+  <%= link_to 'Back', admin_activity_types_path, class: 'btn' %>
 </div>
