+++ app/views/shared/_manage_menu.html.erb
@@ -2,7 +2,7 @@
   <a class="nav-link dropdown-toggle" data-toggle="dropdown" href="#" role="button" aria-haspopup="true" aria-expanded="false">Manage</a>
   <div class="dropdown-menu">
     <%= link_to "Activity Categories", admin_activity_categories_path, class: 'dropdown-item' %>
-    <%= link_to "Activity Types", activity_types_path, class: 'dropdown-item' %>
+    <%= link_to "Activity Types", admin_activity_types_path, class: 'dropdown-item' %>
     <%= link_to "Programme Types", admin_programme_types_path, class: 'dropdown-item' %>
     <%= link_to "Alert Types", admin_alert_types_path, class: 'dropdown-item' %>
     <%= link_to "Calendars", admin_calendars_path, class: 'dropdown-item' %>
